% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigenMapMatMult.R
\name{eigenMapMatMult}
\alias{eigenMapMatMult}
\title{Wrapper for eigenMapMatMult C++ function}
\usage{
eigenMapMatMult(A, B)
}
\arguments{
\item{A}{Matrix A}

\item{B}{Matrix B}
}
\value{
Matrix product A \%*\% B
}
\description{
Wrapper for eigenMapMatMult C++ function
}
